+++ gulpfile.js
@@ -1,19 +1,11 @@
 // #####################################################################################################################
 // #IMPORTS#
 const gulp = require('gulp');
-const gutil = require('gulp-util');
-const gulpif = require('gulp-if');
-const sourcemaps = require('gulp-sourcemaps');
+const log = require('fancy-log');
+const PluginError = require('plugin-error');
 const webpack = require('webpack');
-const postcss = require('gulp-postcss');
-const sass = require('gulp-sass');
-const cleanCSS = require('gulp-clean-css');
-const autoprefixer = require('autoprefixer');
-const flexbugs = require('postcss-flexbugs-fixes');
-const initial = require('postcss-initial');
-const watch = require('gulp-watch');
 
-var argv = require('minimist')(process.argv.slice(2)); // eslint-disable-line
+const argv = require('minimist')(process.argv.slice(2)); // eslint-disable-line
 
 // #####################################################################################################################
 // #SETTINGS#
@@ -23,13 +15,6 @@ var options = {
 var PROJECT_ROOT = __dirname + '/djangocms_alias/static/djangocms_alias';
 var PROJECT_PATH = {
     js: PROJECT_ROOT + '/js',
-    sass: PROJECT_ROOT + '/sass',
-    css: PROJECT_ROOT + '/css',
-};
-
-var PROJECT_PATTERNS = {
-    js: [PROJECT_PATH.js + '/*.js', '!' + PROJECT_PATH.js + '/dist/*.js'],
-    sass: [PROJECT_PATH.sass + '/**/*.{scss,sass}'],
 };
 
 var webpackBundle = function(opts) {
@@ -43,9 +28,9 @@ var webpackBundle = function(opts) {
 
         webpack(config, function(err, stats) {
             if (err) {
-                throw new gutil.PluginError('webpack', err);
+                throw new PluginError('webpack', err);
             }
-            gutil.log('[webpack]', stats.toString({ maxModules: Infinity, colors: true, optimizationBailout: true }));
+            log('[webpack]', stats.toString({ maxModules: Infinity, colors: true, optimizationBailout: true }));
             if (typeof done !== 'undefined' && (!opts || !opts.watch)) {
                 done();
             }
@@ -55,46 +40,3 @@ var webpackBundle = function(opts) {
 
 gulp.task('bundle:watch', webpackBundle({ watch: true }));
 gulp.task('bundle', webpackBundle());
-
-gulp.task('sass', function () {
-    return gulp.src(PROJECT_PATTERNS.sass)
-        .pipe(gulpif(argv.debug, sourcemaps.init()))
-        .pipe(sass())
-        .on('error', function (error) {
-            gutil.log(gutil.colors.red(
-                'Error (' + error.plugin + '): ' + error.messageFormatted)
-            );
-
-            if (process.env.EXIT_ON_ERRORS) {
-                process.exit(1); // eslint-disable-line
-            } else {
-                // in dev mode - just continue
-                this.emit('end');
-            }
-        })
-        .pipe(
-            postcss([
-                initial,
-                autoprefixer({
-                    // browsers are coming from browserslist file
-                    cascade: false,
-                }),
-                flexbugs,
-            ])
-        )
-        .pipe(gulpif(!argv.debug, cleanCSS({
-            rebase: false,
-        })))
-        .pipe(gulpif(argv.debug, sourcemaps.write()))
-        .pipe(gulp.dest(PROJECT_PATH.css));
-});
-
-gulp.task('watch', function() {
-    gulp.start('bundle:watch');
-
-    watch(PROJECT_PATTERNS.sass, function () {
-        return gulp.start('sass');
-    });
-});
-
-gulp.task('default', ['watch']);